# [Java] Speedup generateNewUUID