### Problem

Current `generageNewUUID` is slow because:
* `new SecureRandom()` created for each generated UUID
* `SecureRandom` is cryptographic secure, but slower than `Random` 4x times
* According to [Java docs](https://docs.oracle.com/javase/8/docs/api/java/util/Random.html): "However, the concurrent use of the same java.util.Random instance across threads may encounter contention and consequent poor performance"

### Solution

* Do not create random generator on each method call, use static field.
* Use `Random` instead of `SecureRandom`, but seed it with SecureRandom output, to avoid returning the same values (due to a pseudo-random nature of Random).
* Wrap random generator with `ThreadLocal<T>` to avoid contention in multithread environment.
* Add tests for generated UUIDs uniqueness, including multithread case.

Benchmarks:

<details>
<summary>OpenLineageUUIDBenchmark.java</summary>

```java
package org.openlineage;

import org.openjdk.jmh.annotations.*;
import org.openjdk.jmh.infra.Blackhole;

import com.fasterxml.uuid.EthernetAddress;
import com.fasterxml.uuid.Generators;
import com.fasterxml.uuid.impl.RandomBasedGenerator;
import com.fasterxml.uuid.impl.TimeBasedGenerator;
import com.fasterxml.uuid.impl.TimeBasedEpochRandomGenerator;

import java.lang.ThreadLocal;
import java.time.Instant;
import java.util.Random;
import java.security.SecureRandom;
import java.util.concurrent.TimeUnit;
import java.util.UUID;

@BenchmarkMode(Mode.AverageTime)
@OutputTimeUnit(TimeUnit.NANOSECONDS)
@State(Scope.Benchmark)
@Warmup(iterations = 3)
@Measurement(iterations = 3)
@Fork(1)
public class OpenLineageUUIDBenchmark {
  private static final Instant NOW = Instant.now();
  private static final SecureRandom SECURE_RANDOM = new SecureRandom();
  private static final Random RANDOM = new Random();
  private static final ThreadLocal<Random> THREAD_LOCAL_RANDOM =
      new ThreadLocal<Random>() {
        @Override
        protected Random initialValue() {
          SecureRandom trueRandom = new SecureRandom();
          long seed = trueRandom.nextLong();
          return new Random(seed);
        }
      };

  public static UUID generateNewUUID_SecureRandom_New() {
    return generateNewUUID_SecureRandom_New(Instant.now());
  }

  public static UUID generateNewUUID_SecureRandom_New(Instant instant) {
    long time = instant.toEpochMilli();
    SecureRandom random = new SecureRandom();
    long msb = (time << 16) | (random.nextLong() & 0x0fffL) | 0x7000L;
    long lsb = (random.nextLong() & 0x3fffffffffffffffL) | 0x8000000000000000L;
    return new UUID(msb, lsb);
  }

  public static UUID generateNewUUID_SecureRandom_Reuse() {
    return generateNewUUID_SecureRandom_Reuse(Instant.now());
  }

  public static UUID generateNewUUID_SecureRandom_Reuse(Instant instant) {
    long time = instant.toEpochMilli();
    long msb = (time << 16) | (SECURE_RANDOM.nextLong() & 0x0fffL) | 0x7000L;
    long lsb = (SECURE_RANDOM.nextLong() & 0x3fffffffffffffffL) | 0x8000000000000000L;
    return new UUID(msb, lsb);
  }

  public static UUID generateNewUUID_Random() {
    return generateNewUUID_Random(Instant.now());
  }

  public static UUID generateNewUUID_Random(Instant instant) {
    long time = instant.toEpochMilli();
    long msb = (time << 16) | (RANDOM.nextLong() & 0x0fffL) | 0x7000L;
    long lsb = (RANDOM.nextLong() & 0x3fffffffffffffffL) | 0x8000000000000000L;
    return new UUID(msb, lsb);
  }

  public static UUID generateNewUUID_Random_ThreadLocal() {
    return generateNewUUID_Random_ThreadLocal(Instant.now());
  }

  public static UUID generateNewUUID_Random_ThreadLocal(Instant instant) {
    long time = instant.toEpochMilli();
    Random random = THREAD_LOCAL_RANDOM.get();
    long msb = (time << 16) | (random.nextLong() & 0x0fffL) | 0x7000L;
    long lsb = (random.nextLong() & 0x3fffffffffffffffL) | 0x8000000000000000L;
    return new UUID(msb, lsb);
  }

  @Benchmark
  public void benchmarkGenerateNewUUID_SecureRandom_New(Blackhole bh) {
      bh.consume(generateNewUUID_SecureRandom_New());
  }

  @Benchmark
  @Threads(4)
  public void benchmarkGenerateNewUUID_SecureRandom_New_Parallel4(Blackhole bh) {
      bh.consume(generateNewUUID_SecureRandom_New());
  }

  @Benchmark
  public void benchmarkGenerateNewUUID_ForInstant_SecureRandom_New(Blackhole bh) {
      bh.consume(generateNewUUID_SecureRandom_New(NOW));
  }

  @Benchmark
  @Threads(4)
  public void benchmarkGenerateNewUUID_ForInstant_SecureRandom_New_Parallel4(Blackhole bh) {
      bh.consume(generateNewUUID_SecureRandom_New(NOW));
  }

  @Benchmark
  public void benchmarkGenerateNewUUID_SecureRandom_Reuse(Blackhole bh) {
      bh.consume(generateNewUUID_SecureRandom_Reuse());
  }

  @Benchmark
  @Threads(4)
  public void benchmarkGenerateNewUUID_SecureRandom_Reuse_Parallel4(Blackhole bh) {
      bh.consume(generateNewUUID_SecureRandom_Reuse());
  }

  @Benchmark
  public void benchmarkGenerateNewUUID_ForInstant_SecureRandom_Reuse(Blackhole bh) {
      bh.consume(generateNewUUID_SecureRandom_Reuse(NOW));
  }

  @Benchmark
  @Threads(4)
  public void benchmarkGenerateNewUUID_ForInstant_SecureRandom_Reuse_Parallel4(Blackhole bh) {
      bh.consume(generateNewUUID_SecureRandom_Reuse(NOW));
  }

  @Benchmark
  public void benchmarkGenerateNewUUID_Random(Blackhole bh) {
      bh.consume(generateNewUUID_Random());
  }

  @Benchmark
  @Threads(4)
  public void benchmarkGenerateNewUUID_Random_Parallel4(Blackhole bh) {
      bh.consume(generateNewUUID_Random());
  }

  @Benchmark
  public void benchmarkGenerateNewUUID_ForInstant_Random(Blackhole bh) {
      bh.consume(generateNewUUID_Random(NOW));
  }

  @Benchmark
  @Threads(4)
  public void benchmarkGenerateNewUUID_ForInstant_Random_Parallel4(Blackhole bh) {
      bh.consume(generateNewUUID_Random(NOW));
  }

  @Benchmark
  public void benchmarkGenerateNewUUID_Random_ThreadLocal(Blackhole bh) {
      bh.consume(generateNewUUID_Random_ThreadLocal());
  }

  @Benchmark
  @Threads(4)
  public void benchmarkGenerateNewUUID_Random_ThreadLocal_Parallel4(Blackhole bh) {
      bh.consume(generateNewUUID_Random_ThreadLocal());
  }

  @Benchmark
  public void benchmarkGenerateNewUUID_ForInstant_ThreadLocal_Random(Blackhole bh) {
      bh.consume(generateNewUUID_Random_ThreadLocal(NOW));
  }

  @Benchmark
  @Threads(4)
  public void benchmarkGenerateNewUUID_ForInstant_ThreadLocal_Random_Parallel4(Blackhole bh) {
      bh.consume(generateNewUUID_Random_ThreadLocal(NOW));
  }
}
```
</details>

```
Benchmark                                                                                  Mode  Cnt     Score      Error  Units

OpenLineageUUIDBenchmark.benchmarkGenerateNewUUID_SecureRandom_New                         avgt    3  3346.101 ± 3013.374  ns/op
OpenLineageUUIDBenchmark.benchmarkGenerateNewUUID_SecureRandom_New_Parallel4               avgt    3  7167.171 ± 7475.353  ns/op
OpenLineageUUIDBenchmark.benchmarkGenerateNewUUID_ForInstant_SecureRandom_New              avgt    3  3497.326 ± 3599.744  ns/op
OpenLineageUUIDBenchmark.benchmarkGenerateNewUUID_ForInstant_SecureRandom_New_Parallel4    avgt    3  6999.427 ± 3907.263  ns/op

OpenLineageUUIDBenchmark.benchmarkGenerateNewUUID_SecureRandom_Reuse                       avgt    3   598.105 ±  236.201  ns/op
OpenLineageUUIDBenchmark.benchmarkGenerateNewUUID_SecureRandom_Reuse_Parallel4             avgt    3  4915.909 ± 1511.498  ns/op
OpenLineageUUIDBenchmark.benchmarkGenerateNewUUID_ForInstant_SecureRandom_Reuse            avgt    3   586.464 ±  298.559  ns/op
OpenLineageUUIDBenchmark.benchmarkGenerateNewUUID_ForInstant_SecureRandom_Reuse_Parallel4  avgt    3  4212.199 ±  689.010  ns/op

OpenLineageUUIDBenchmark.benchmarkGenerateNewUUID_Random                                   avgt    3    72.842 ±   39.020  ns/op
OpenLineageUUIDBenchmark.benchmarkGenerateNewUUID_Random_Parallel4                         avgt    3  1441.737 ± 1408.846  ns/op
OpenLineageUUIDBenchmark.benchmarkGenerateNewUUID_ForInstant_Random                        avgt    3    37.866 ±   67.863  ns/op
OpenLineageUUIDBenchmark.benchmarkGenerateNewUUID_ForInstant_Random_Parallel4              avgt    3  1327.813 ±  263.878  ns/op

OpenLineageUUIDBenchmark.benchmarkGenerateNewUUID_Random_ThreadLocal                       avgt    3    75.682 ±   49.752  ns/op
OpenLineageUUIDBenchmark.benchmarkGenerateNewUUID_Random_ThreadLocal_Parallel4             avgt    3   104.993 ±   28.207  ns/op
OpenLineageUUIDBenchmark.benchmarkGenerateNewUUID_ForInstant_ThreadLocal_Random            avgt    3    42.778 ±   35.800  ns/op
OpenLineageUUIDBenchmark.benchmarkGenerateNewUUID_ForInstant_ThreadLocal_Random_Parallel4  avgt    3    57.275 ±   14.002  ns/op
```

#### One-line summary:

[Java] Speedup generateNewUUID

### Checklist

- [X] You've [signed-off](https://github.com/OpenLineage/OpenLineage/blob/main/why-the-dco.md) your work
- [X] Your pull request title follows our [guidelines](https://github.com/OpenLineage/OpenLineage/blob/main/CONTRIBUTING.md#creating-pull-requests)
- [X] Your changes are accompanied by tests (_if relevant_)
- [X] Your change contains a [small diff](https://kurtisnusbaum.medium.com/stacked-diffs-keeping-phabricator-diffs-small-d9964f4dcfa6) and is self-contained
- [ ] You've updated any relevant documentation (_if relevant_)
- [X] Your comment includes a one-liner for the changelog about the specific purpose of the change (_not required for changes to tests, docs, or CI config_)
- [ ] You've versioned the core OpenLineage model or facets according to [SchemaVer](https://docs.snowplowanalytics.com/docs/pipeline-components-and-applications/iglu/common-architecture/schemaver) (_if relevant_)
- [ ] You've added a [header](https://github.com/OpenLineage/OpenLineage/tree/main/.github/header_templates.md) to source files (_if relevant_)

----
SPDX-License-Identifier: Apache-2.0\
Copyright 2018-2025 contributors to the OpenLineage project